Move operators out of skunk.sharp.where (rename → skunk.sharp.ops)
Comparison / LIKE / IN / IS NULL operators produce a plain
`TypedExpr[Boolean]`. Postgres accepts boolean expressions anywhere an
expression is valid — WHERE, SELECT projections, HAVING, ORDER BY,
function arguments, CASE WHEN predicates — so burying them in a package
called `where` was misleading.

Moved `skunk.sharp.where.Ops` → `skunk.sharp.ops.ExprOps`. The
`skunk.sharp.where` package keeps only the `Where` type alias and the
logical combinators (`&&`, `||`, unary `!`, `and`, `or`, `not`), which
genuinely are about composing booleans downstream. Internal imports
flipped (`Shared` / `PgNull` / `ArrayOps` / `dsl.package` re-exports,
and two test files that did `import skunk.sharp.where.*`). User-facing
surface via `import skunk.sharp.dsl.*` unchanged.

Also added three `SelectSuite` tests exercising the operators in a
projection position — `users.select(u => u.age >= 18)` renders
`SELECT "age" >= $1 FROM "users"`, and the tuple form
`(u.email, u.age >= 18)` decodes to `(String, Boolean)`. Also
`u.deleted_at.isNull` in projection — the same nullable-gating as in
WHERE applies.

ORDER BY on a boolean expression would work in Postgres but our `.asc`
/ `.desc` only extend `TypedColumn`, not general `TypedExpr` — noted in
a comment as a separate gap.

Tests
  - 224 core (up 3 from the new projection tests).
  - 99 integration.
  - CI-mode clean, scalafmt clean.
